requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
Pillow==9.4.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0
django-request~=1.6.2
drf_yasg==1.21.5

